+++ src/FlaUILibrary/flaui/module/application.py
@@ -24,7 +24,7 @@ class Application(ModuleInterface):
         """Application module wrapper for FlaUI usage.
 
         Args:
-            automation (Object): UIA3 automation object from FlaUI.
+            automation (Object): UIA3/UIA2 automation object from FlaUI.
         """
         self._application = None
         self._automation = automation